requirements.txt
Adafruit-Blinka==8.44.4
adafruit-circuitpython-bme280==2.6.24
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.71.0
Adafruit-PureIO==1.1.11
adafruit-python-shell==1.8.1
args==0.1.0
binho-host-adapter==0.1.6
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
contourpy==1.2.1
cycler==0.12.1
docutils==0.21.2
fonttools==4.53.1
idna==3.7
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyftdi==0.55.4
Pygments==2.18.0
pyparsing==3.1.2
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyusb==1.2.1
requests==2.32.3
rpi-backlight==2.6.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
schedule==1.2.2
shiboken6==6.7.2
six==1.16.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
